+++ packages/solospec/src/theme/themes/bikeshed/components.tsx
@@ -1,20 +1,21 @@
 import type { ThemeSlots, RenderPageVm } from '#src/theme/types';
-import type { DocumentMetadata, TocEntry } from '@openuji/speculator';
+import type { TocEntry, Preprocess } from '@openuji/speculator';
 import { baseSlots, baseComponents } from '#src/theme/themes/base/components';
 import { AstComponentsContext } from '#src/theme/themes/base/context';
 // import { TocNav } from '#src/theme/themes/bikeshed/TocNav';
 
 import type { JSX } from 'preact';
-
-function formatPersonList(people: DocumentMetadata['editors'] | DocumentMetadata['authors']): JSX.Element | null {
+type PersonEntry = Preprocess.PersonEntry;
+function formatPersonList(people: PersonEntry[]): JSX.Element | null {
   if (!people || people.length === 0) return null;
+
   return (
     <>
       {people.map((person, i) => (
         <span key={i}>
           {i > 0 ? <br /> : null}
-          {person.url ? <a href={person.url}>{person.name}</a> : person.name}
-          {person.company ? ` (${person.company})` : null}
+          {person.url ? <a href={person.url}>{person.name}</a> : (person.email ? <a href={`mailto:${person.email}`}>{person.name}</a> : person.name) }
+          {person.company ? person.companyUrl ? <> (<a href={person.companyUrl}>{person.company}</a>)</> : <>(person.company)</> : null}
         </span>
       ))}
     </>
@@ -140,16 +141,16 @@ function BikeshedHeader(props: Parameters<ThemeSlots['Header']>[0]) {
             </>
           ) : null}
 
-          {custom.formerEditors && Array.isArray(custom.formerEditors) && custom.formerEditors.length > 0 ? (
+          {custom.formereditor && Array.isArray(custom.formereditor) && custom.formereditor.length > 0 ? (
             <>
               <dt>Former Editors:</dt>
-              <dd>{formatPersonList(custom.formerEditors as DocumentMetadata['editors'])}</dd>
+              <dd>{formatPersonList(custom.formereditor)}</dd>
             </>
           ) : null}
         </dl>
       </details>
-      {meta.copyright ? (() => {
-        const copyright = String(meta.copyright)
+      {custom.copyright ? (() => {
+        const copyright = String(custom.copyright)
           .replace(/\[YEAR\]/g, String(new Date().getFullYear()))
           .replace(/\[TITLE\]/g, vm.titleText);
         return (